Changed directory/file names
